rep,nanoduration-method {nanotime}	R Documentation

Replicate Elements

Description

Replicates the values in 'x' similarly to the default method.

Usage

## S4 method for signature 'nanoduration'
rep(x, ...)

Arguments

x
a vector of nanoduration

...
further arguments:

'times' an integer-valued vector giving the (non-negative) number of times to repeat each element if of length 'length(x)', or to repeat the whole vector if of length 1. Negative or 'NA' values are an error. A 'double' vector is accepted, other inputs being coerced to an integer or double vector.

'length.out' non-negative integer. The desired length of the output vector. Other inputs will be coerced to a double vector and the first element taken. Ignored if 'NA' or invalid.

'each' non-negative integer. Each element of 'x' is repeated 'each' times. Other inputs will be coerced to an integer or double vector and the first element taken. Treated as '1' if 'NA' or invalid.

[Package nanotime version 0.3.13 Index]